+++ tests/testthat/test-tidyeval.R
@@ -5,6 +5,12 @@ test_that("simple expressions left as is", {
   expect_equal(capture_dot(dt, 10), 10)
   expect_equal(capture_dot(dt, x), quote(x))
   expect_equal(capture_dot(dt, x + y), quote(x + y))
+
+  # logicals
+  expect_equal(eval(capture_dot(dt, T), globalenv()), TRUE)
+  expect_equal(eval(capture_dot(dt, F), globalenv()), FALSE)
+  expect_equal(capture_dot(dt, TRUE), TRUE)
+  expect_equal(capture_dot(dt, FALSE), FALSE)
 })
 
 test_that("existing non-variables get inlined", {
@@ -13,9 +19,6 @@ test_that("existing non-variables get inlined", {
   n <- 10
   expect_equal(capture_dot(dt, x + n), quote(x + 10))
   expect_equal(capture_dot(dt, x + m), quote(x + m))
-
-  # except when not in j
-  expect_equal(capture_dot(dt, x + n, j = FALSE), quote(x + n))
 })
 
 test_that("unless we're operating in the global environment", {
@@ -32,6 +35,7 @@ test_that("using environment of inlined quosures", {
   quo <- new_quosure(quote(x + n), env(n = 20))
 
   expect_equal(capture_dot(dt, f(!!quo)), quote(f(x + 20)))
+  expect_equal(capture_dot(dt, f(!!quo), j = FALSE), quote(f(x + 20)))
 })
 
 test_that(". gets converted to .SD", {